환전 시 my page에 ETP 사용 내역 렌더링 되도록 처리

+++ server/src/main/java/com/ict/extravel/domain/pointexchange/dto/response/PayConfirmResponseDTO.java
@@ -20,9 +20,9 @@ public class PayConfirmResponseDTO {
 
     private Integer memberId;
 
-    private Integer amount;
+    private BigDecimal amount;
 
-    private Float plusPoint;
+    private BigDecimal plusPoint;
 
     private LocalDateTime createdAt;
